Change the type for the statistics counter.
Signed int could overflow for the Cornacchia step with a 25000 digit
number.

* lib/cm-impl.h (cm_stat_t): Replace int by unsigned long int.
* lib/ecpp.c (ecpp1), lib/file.c (write_stat, read_stat): Adapt the code
  for file input and output.

+++ lib/cm-impl.h
@@ -61,7 +61,7 @@ typedef struct {
 
 
 typedef struct {
-   int counter [4];
+   unsigned long int counter [4];
    cm_timer_t timer [8];
 }
 __cm_stat_struct;
